# src/main/java/com/example/orderpayment/dto/PaymentSnapshot.java
package com.example.orderpayment.dto;

public record PaymentSnapshot(
        String status,
        long authorizedAmount,
        long capturedAmount,
        long refundedAmount,
        long version) {}