# src/main/java/com/example/orderpayment/dto/ReconciliationResponse.java
package com.example.orderpayment.dto;

public record ReconciliationResponse(
        String date,
        String restaurantId,
        String restaurantTimezone,
        long ordersCreated,
        long authorizedAmount,
        long capturedAmount,
        long refundedAmount,
        long netCollected,
        long failedAuthCount,
        long failedCaptureCount) {}